Add xcuserdata to gitignore
.gitignore files are a way of telling Git to not worry about committing
certain files to the repository.  In this case, xcuserdata is a folder
that XCode uses to keep track of stuff like positions of windows, how
far you are zoomed in on the storyboard, etc.  Stuff that has nothing to
do with the actual project that other people would need to keep track
of.  If you're working by youself it's not a problem, but when you're
working with other people you'll typically agree on a set of things that
won't get committed.  Everything that is needed to actually build and
run the project still needs to be committed though!