factor out myWeatherRange
the ABCDE controllers all have almost exactly the same viewDidLoad.
Extracting the differences into individual methods will allow us to
create a common controller that shares most of the same behavior - this
is an important principle called DRY (short for "don't repeat
yourself").  In general, you should only be doing something once in a
codebase.  If you find yourself needing to do it in multiple places,
you're going to want to extract it to a utility function or a common
superclass.  This commit on its own doesn't change any behavior, but
you'll see in a little while that it makes it easier to change the
behavior of the whole app when we make it so that we only download the
weather once.