extract myDayOffset
This one is probably going to seem even more confusing, but it's towards
the same goal of making the ABCDE controllers more uniform.  I added a
dateLbl outlet to both A and B, even though they don't show a date.
What this allowed me to do, though, is have exactly the same getDay func
for all five controllers.  The only differences between the five
controllers are: how many days from today is the date they
display (which I extracted into the myDayOffset method), and whether
they actually update a label with the date.  For A and B, since the
outlet is not hooked up in Interface Builder, that outlet will be
'nil' (which is a special value that means "nothing").  We can test and
see that there's no label for those controllers and we won't run the
update logic.  Again, this will make more sense why we're doing this in
a second :)