update tempLbl
Ah, I forgot one difference, the A controller has a temperature label.
Did the same thing as with the dateLbl, I added tempLbl outlets to all
controllers and tested to see if it was nil before updating it.  Now all
of ABCDE have the same viewDidLoad (except for some variable names), so
we can extract a superclass in the next commit.