use myWeatherRange to determine weather image
Ack, one more change I missed before we can factor out the superclass.
The weather image comes from an entry in the list which is currently
hardcoded per day.  It's the first weather entry for that day, so we can
just use the myWeatherRange func that we already extracted to find that.