extract superclass
So this is a big commit - this makes all the ABCDE controllers descend
from a new controller, the DayViewController.  DayViewController
basically has the exact same stuff as everything else, it just doesn't
know what day it refers to, so it relies on the child classes to fill in
the myDayOffset and the myWeatherRange.  This is called "inheritance" -
you'll see that AViewController now has "DayViewController" listed at
the top instead of UIViewController and PageDelegate.  When you read
that, you should read it like "An AViewController is a DayViewController
except for the following changes in the braces".  Now that they're all
mostly the same, we can do this.