extract updateWeather method
I basically just moved everything that happens after we download the
weather into its own func.  This will allow us to move where the weather
is downloaded very easily.  Notice that now we have the superclass, we
only had to make this big change in one class and not have to worry
about copy/paste errors!