move weather-downloading to ViewController
So this is the main change that lets the weather be downloaded only
once.  There are a couple changes in here that go along with it.  First,
the ViewController needs a weather variable that it can keep the
downloaded weather in.  It also needs to pass that weather to the
DayViewControllers when creating them.  I added the updateWeather func
to the PageDelegate protocol so that ViewController can call to it.
And (this is sort of a tricky bit) - since pages can load before the
download is complete, once the download completes, we need to go through
all of our existing pages and let them know about the weather.

So that's it - ViewController downloads the weather once when it loads,
and then parcels it out to the individual view controllers!